+++ adcode/admin.py
@@ -10,8 +10,11 @@ class PlacementInline(admin.StackedInline):
 
 
 class SectionAdmin(admin.ModelAdmin):
+    list_display = ('name', 'pattern', 'priority', )
+    list_editable = ('priority', )
     inlines = (PlacementInline, )
     prepopulated_fields = {'slug': ('name', )}
+    ordering = ('-priority', 'name', )
    
 
 class SizeAdmin(admin.ModelAdmin):
@@ -23,6 +26,7 @@ class PlacementAdmin(admin.ModelAdmin):
     list_display = ('name', 'size', )
     list_filter = ('size', 'sections', )
     filter_horizontal = ('sections', )
+    prepopulated_fields = {'slug': ('name', )}
 
 
 admin.site.register(Section, SectionAdmin)