+++ adcode/tests/__init__.py
@@ -1,3 +1,5 @@
+from .base import FixturesTestCase
 from .context_processors import CurrentPlacementsTestCase
+from .performance import QueryCountsTestCase
 from .templatetags import RenderHeaderTestCase, RenderPlacementTestCase
 from .validators import PatternValidatorTestCase