+++ adcode/tests/context_processors.py
@@ -41,7 +41,7 @@ class CurrentPlacementsTestCase(AdCodeDataTestCase):
         placement.sections.add(section)
         context = current_placements(request)
         placements = context[PLACEMENTS_CONTEXT_KEY]
-        self.assertEqual(placements.count(), 1)
+        self.assertEqual(len(placements), 1)
         self.assertTrue(placement in placements)
 
     def test_match_url_no_placements(self):
@@ -50,4 +50,12 @@ class CurrentPlacementsTestCase(AdCodeDataTestCase):
         section = self.create_section(pattern='^/foo/')
         context = current_placements(request)
         placements = context[PLACEMENTS_CONTEXT_KEY]
-        self.assertEqual(placements.count(), 0)
+        self.assertEqual(len(placements), 0)
+
+    def test_match_priority(self):
+        "Resolve overlapping regex patterns by using priority flag."
+        request = self.factory.get('/foo/bar')
+        section = self.create_section(pattern='^/foo/', priority=0)
+        other_section = self.create_section(pattern='^/foo/bar', priority=1)
+        context = current_placements(request)
+        self.assertEqual(context[SECTION_CONTEXT_KEY], other_section)